ada22-main/user-ms/requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asyncpg==0.25.0
fastapi==0.75.0
idna==3.3
iso8601==0.1.16
pydantic==1.9.0
pypika-tortoise==0.1.3
pytz==2022.1
sniffio==1.2.0
starlette==0.17.1
tortoise-orm==0.18.1
